chore: create singleton for primitive mapping
