+++ src/Blazor.SourceGenerators/CSharp/CSharpProperty.cs
@@ -4,7 +4,7 @@
 namespace Blazor.SourceGenerators.CSharp;
 
 /// <summary>
-/// A record the represents various C# members, such as properties, delegates and events.
+/// A record that represents various C# members, such as properties, delegates, and events.
 /// </summary>
 internal record CSharpProperty(
     string RawName,
@@ -12,34 +12,38 @@ internal record CSharpProperty(
     bool IsNullable = false,
     bool IsReadonly = false) : CSharpType(RawName, RawTypeName, IsNullable)
 {
-    public string MappedTypeName
-    {
-        get
-        {
-            var mappedTypeName = TypeMap.PrimitiveTypes[RawTypeName];
-            if (mappedTypeName == RawTypeName)
-            {
-                if (IsArray)
-                {
-                    mappedTypeName = mappedTypeName
-                        .Replace("[]", "")
-                        .Replace("ReadonlyArray<", "")
-                        .Replace(">", "");
-                }
+    /// <summary>
+    /// Gets the mapped type name, resolving primitive types and handling arrays and nullability.
+    /// </summary>
+    public string MappedTypeName => GetMappedTypeName();
 
-                if (IsNullable)
-                {
-                    mappedTypeName = mappedTypeName
-                        .Replace("| null", "");
-                }
-            }
-
-            return mappedTypeName;
-        }
-    }
-
+    /// <summary>
+    /// Determines if the property is an indexer.
+    /// </summary>
     public bool IsIndexer => RawName.StartsWith("[") && RawName.EndsWith("]");
 
-    public bool IsArray => RawTypeName.EndsWith("[]") ||
-        (RawTypeName.StartsWith("ReadonlyArray<") && RawTypeName.EndsWith(">"));
+    /// <summary>
+    /// Determines if the property is an array.
+    /// </summary>
+    public bool IsArray => RawTypeName.EndsWith("[]") || (RawTypeName.StartsWith("ReadonlyArray<") && RawTypeName.EndsWith(">"));
+
+    private string GetMappedTypeName()
+    {
+        var mappedTypeName = Primitives.Instance[RawTypeName];
+
+        if (IsArray)
+        {
+            mappedTypeName = mappedTypeName
+                .Replace("[]", "")
+                .Replace("ReadonlyArray<", "")
+                .Replace(">", "");
+        }
+
+        if (IsNullable)
+        {
+            mappedTypeName = mappedTypeName.Replace("| null", "");
+        }
+
+        return mappedTypeName;
+    }
 }
